Move useSaveFrontstageSettings to separate module.
