+++ ui/appui-react/src/test/widget-panels/Frontstage.test.tsx
@@ -6,8 +6,7 @@ import { Logger } from "@itwin/core-bentley";
 import { IModelApp } from "@itwin/core-frontend";
 import type { UiStateStorageResult } from "@itwin/core-react";
 import { Size, UiStateStorageStatus } from "@itwin/core-react";
-import { render, screen } from "@testing-library/react";
-import { act, renderHook } from "@testing-library/react-hooks";
+import { render, renderHook, screen } from "@testing-library/react";
 import produce from "immer";
 import * as React from "react";
 import { Provider } from "react-redux";
